elsevier_spider: improvements
* Moves `range_as_string` to utils.py and adds tests.

* Makes `parse_node` a bit more clear.
